+++ LookupAnything/Framework/Fields/Models/RecipeItemEntry.cs
@@ -6,5 +6,6 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Fields.Models
     /// <param name="Sprite">The sprite to display.</param>
     /// <param name="DisplayText">The display text for the item name and count.</param>
     /// <param name="Quality">The item quality that will be produced, if applicable.</param>
-    internal record RecipeItemEntry(SpriteInfo? Sprite, string DisplayText, int? Quality);
+    /// <param name="IsGoldPrice">Whether this is a gold price, rather than an ingredient.</param>
+    internal record RecipeItemEntry(SpriteInfo? Sprite, string DisplayText, int? Quality, bool IsGoldPrice);
 }